+++ src-tauri/src/lib.rs
@@ -1,7 +1,3 @@
-// Prevents additional console window on Windows in release, DO NOT REMOVE!!
-#![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
-#![macro_use]
-
 mod helpers;
 mod translator;
 mod types;
@@ -9,7 +5,7 @@ mod types;
 use tauri_specta::{collect_commands, Builder};
 
 use crate::translator::file_system::{
-    create_translation, get_xliff_meta, open_translation, open_xliff, save_translation,
+    create_translation, get_xliff_meta, open_translation, open_xliff, save_translation, detect_format
 };
 
 #[cfg_attr(mobile, tauri::mobile_entry_point)]
@@ -19,7 +15,8 @@ pub fn run() {
         open_translation,
         save_translation,
         get_xliff_meta,
-        open_xliff
+        open_xliff,
+        detect_format
     ]);
     tauri::Builder::default()
         .plugin(tauri_plugin_updater::Builder::new().build())