+++ src/loaders/settings/file-formats.tsx
@@ -6,10 +6,10 @@ export default function FileFormatSettings(){
                <SettingsItemLoader>
                     <SettingsOptionLoader optionType="selector" width={128}/>
                </SettingsItemLoader>
-               {/* <SettingsItemLoader>
+               <SettingsItemLoader>
                     <SettingsOptionLoader optionType="switch"/>
                     <SettingsOptionLoader optionType="switch"/>
-               </SettingsItemLoader> */}
+               </SettingsItemLoader>
                <SettingsItemLoader>
                     <SettingsOptionLoader optionType="switch"/>
                     <SettingsOptionLoader optionType="switch"/>